requirements.txt
Flask==3.1.0
flask_cors==5.0.1
huggingface_hub==0.30.2
Pillow==11.2.1
pyttsx3==2.98
speechrecognition==3.14.2
torch==2.6.0
transformers==4.51.1
